:truck: [cookiecutter] Move iterpaths to main
